refactor: refactor data files
