+++ simulations/vip-665/utils/remoteSim.ts
@@ -6,10 +6,10 @@ import { NETWORK_ADDRESSES } from "src/networkAddresses";
 import { expectEvents, initMainnetUser } from "src/utils";
 import { forking, testForkedNetworkVipCommands } from "src/vip-framework";
 
-import vip665 from "../../../vips/vip-665/bscmainnet";
-import { ZERO } from "../../../vips/vip-665/data/bnb";
-import { EXPECTED_ROLE_EVENTS, REMOTE_ACM, remoteRowsFor } from "../../../vips/vip-665/data/remote";
-import { SYNC_CASH_MARKETS, SYNC_CASH_SIG } from "../../../vips/vip-665/data/syncCash";
+import vip665, { EXPECTED_ROLE_EVENTS } from "../../../vips/vip-665/bscmainnet";
+import { remoteRowsFor } from "../../../vips/vip-665/data/actionPlan";
+import { REMOTE_ACM, SYNC_CASH_MARKETS, ZERO } from "../../../vips/vip-665/data/addresses";
+import { REDUNDANT_REVOKES, SYNC_CASH_SIG } from "../../../vips/vip-665/data/cleanup";
 import { AGGREGATOR, Chain, grantPermissions, revokePermissions } from "../../../vips/vip-665/utils/commands";
 import { seedAggregator } from "../../../vips/vip-665/utils/seed";
 import ACCESS_CONTROL_MANAGER_ABI from "../abi/AccessControlManager.json";
@@ -21,7 +21,7 @@ const role = (at: string, fn: string) => ethers.utils.solidityKeccak256(["addres
 // RoleGranted / RoleRevoked are emitted by every ACM version. On remotes the grants are the wrapper
 // grantRole(DEFAULT_ADMIN_ROLE) plus the single wildcard syncCash() grant; RoleRevoked = the Critical
 // revokes + the per-market syncCash() revokes + the wrapper revokeRole. Expected counts come from the
-// hardcoded EXPECTED_ROLE_EVENTS oracle in data/remote.ts.
+// hardcoded EXPECTED_ROLE_EVENTS oracle.
 
 // Shared body for the 7 remote-chain simulations. Run each with its matching --fork.
 export const runRemoteSim = (chain: RemoteChain, forkBlock: number) => {
@@ -37,10 +37,17 @@ export const runRemoteSim = (chain: RemoteChain, forkBlock: number) => {
     });
 
     const rows = remoteRowsFor(chain);
+    const redundant = REDUNDANT_REVOKES[chain];
 
     describe(`VIP-665 pre-execution permissions (${chain})`, () => {
       it("the plan has at least one revoke row for this chain", () => {
         expect(rows.length, `revoke rows for ${chain}`).to.be.greaterThan(0);
+      });
+
+      it("redundant: each target-specific grant is currently held by its account", async () => {
+        const acm = new Contract(REMOTE_ACM[chain], ACCESS_CONTROL_MANAGER_ABI, ethers.provider);
+        for (const r of redundant)
+          expect(await acm.hasRole(role(r.contract, r.signature), r.account), `pre redundant ${r.note}`).to.be.true;
       });
 
       it("revoke: Critical currently holds every permission to be revoked", async () => {
@@ -99,6 +106,21 @@ export const runRemoteSim = (chain: RemoteChain, forkBlock: number) => {
             await acm.hasRole(role(row.target, row.signature), guardian),
             `post guardian ${row.signature}@${row.target}`,
           ).to.equal(row.guardian);
+      });
+
+      it("redundant: each target-specific grant was revoked", async () => {
+        const acm = new Contract(REMOTE_ACM[chain], ACCESS_CONTROL_MANAGER_ABI, ethers.provider);
+        for (const r of redundant)
+          expect(await acm.hasRole(role(r.contract, r.signature), r.account), `post redundant ${r.note}`).to.be.false;
+      });
+
+      it("redundant behavioral: each account can still call the function via the surviving wildcard", async () => {
+        const acm = new Contract(REMOTE_ACM[chain], ACCESS_CONTROL_MANAGER_ABI, ethers.provider);
+        for (const r of redundant)
+          expect(
+            await acm.isAllowedToCall(r.account, r.signature, { from: r.contract }),
+            `post callable ${r.note} (${r.signature})`,
+          ).to.be.true;
       });
 
       it("behavioral: Critical is no longer allowed to call any revoked function", async () => {
