--- vips/vip-665/data/syncCash.ts
@@ -1,120 +1,0 @@
-import { NETWORK_ADDRESSES } from "src/networkAddresses";
-
-import type { Permission } from "../utils/commands";
-import { RemoteChain } from "./remote";
-
-export const ZERO = "0x0000000000000000000000000000000000000000";
-export const SYNC_CASH_SIG = "syncCash()";
-
-// Markets whose syncCash() role is currently granted to the NormalTimelock, per chain (verified on-chain).
-export const SYNC_CASH_MARKETS: Record<RemoteChain, string[]> = {
-  ethereum: [
-    "0x0792b9c60C728C1D2Fd6665b3D7A08762a9b28e0",
-    "0x0Ec5488e4F8f319213a14cab188E01fB8517Faa8",
-    "0x0c6B19287999f1e31a5c0a44393b24B62D2C0468",
-    "0x13eB80FDBe5C5f4a7039728E258A6f05fb3B912b",
-    "0x17142a05fe678e9584FA1d88EfAC1bF181bF7ABe",
-    "0x17C07e0c232f2f80DfDbd7a95b942D893A4C5ACb",
-    "0x256AdDBe0a387c98f487e44b85c29eb983413c5e",
-    "0x25C20e6e110A1cE3FEbaCC8b7E48368c7b2F0C91",
-    "0x2d499800239C4CD3012473Cb1EAE33562F0A6933",
-    "0x30aD10Bd5Be62CAb37863C2BfcC6E8fb4fD85BDa",
-    "0x325cEB02fe1C2fF816A83a5770eA0E88e2faEcF2",
-    "0x475d0C68a8CD275c15D1F01F4f291804E445F677",
-    "0x4a240F0ee138697726C8a3E43eFE6Ac3593432CB",
-    "0x4fAfbDc4F2a9876Bd1764827b26fb8dc4FD1dB95",
-    "0x520d67226Bc904aC122dcE66ed2f8f61AA1ED764",
-    "0x5e35C312862d53FD566737892aDCf010cb4928F7",
-    "0x62D9E2010Cff87Bae05B91d5E04605ef864ABc3B",
-    "0x672208C10aaAA2F9A6719F449C4C8227bc0BC202",
-    "0x76697f8eaeA4bE01C678376aAb97498Ee8f80D5C",
-    "0x7c8ff7d2A1372433726f879BD945fFb250B94c65",
-    "0x8716554364f20BCA783cb2BAA744d39361fd1D8d",
-    "0x8C3e3821259B82fFb32B2450A95d2dcbf161C24E",
-    "0xA854D35664c658280fFf27B6eDC6C4195c3229B3",
-    "0xCca202a95E8096315E3F19E46e19E1b326634889",
-    "0xDB6C345f864883a8F4cae87852Ac342589E76D1B",
-    "0xE0ee5dDeBFe0abe0a4Af50299D68b74Cec31668e",
-    "0xE36Ae842DbbD7aE372ebA02C8239cd431cC063d6",
-    "0xEF26C64bC06A8dE4CA5D31f119835f9A1d9433b9",
-    "0xF9E9Fe17C00a8B96a8ac20c4E344C8688D7b947E",
-    "0xa0EE2bAA024cC3AA1BC9395522D07B7970Ca75b3",
-    "0xa836ce315b7A6Bb19397Ee996551659B1D92298e",
-    "0xa8e7f9473635a5CB79646f14356a9Fc394CA111A",
-    "0xb4933AF59868986316Ed37fa865C829Eba2df0C7",
-    "0xba3916302cBA4aBcB51a01e706fC6051AaF272A0",
-    "0xc42E4bfb996ED35235bda505430cBE404Eb49F77",
-    "0xc82780Db1257C788F262FBbDA960B3706Dfdcaf2",
-    "0xd8AdD9B41D4E1cd64Edad8722AB0bA8D35536657",
-    "0xf87c0a64dc3a8622D6c63265FA29137788163879",
-  ],
-  arbitrumone: [
-    "0x246a35E79a3a0618535A469aDaF5091cAA9f7E88",
-    "0x39D6d13Ea59548637104E40e729E4aABE27FE106",
-    "0x4f3a73f318C5EA67A86eaaCE24309F29f89900dF",
-    "0x68a34332983f4Bf866768DD6D6E638b02eF5e1f0",
-    "0x7D8609f8da70fF9027E9bc5229Af4F6727662707",
-    "0x9bb8cEc9C0d46F53b4f2173BB2A0221F66c353cC",
-    "0x9df6B5132135f14719696bBAe3C54BAb272fDb16",
-    "0xAeB0FEd69354f34831fe1D16475D9A83ddaCaDA6",
-    "0xB9F9117d4200dC296F9AcD1e8bE1937df834a2fD",
-    "0xaDa57840B372D4c28623E87FC175dE8490792811",
-  ],
-  basemainnet: [
-    "0x133d3BCD77158D125B75A17Cb517fFD4B4BE64C5",
-    "0x3cb752d175740043Ec463673094e06ACDa2F9a2e",
-    "0x75201D81B3B0b9D17b179118837Be37f64fc4930",
-    "0x7bBd1005bB24Ec84705b04e1f2DfcCad533b6D72",
-    "0xEB8A79bD44cF4500943bf94a2b4434c95C008599",
-  ],
-  zksyncmainnet: [
-    "0x03CAd66259f7F34EE075f8B62D133563D249eDa4",
-    "0x183dE3C349fCf546aAe925E1c7F364EA6FB4033c",
-    "0x1Fa916C27c7C2c4602124A14C77Dbb40a5FF1BE8",
-    "0x1aF23bD57c62A99C59aD48236553D0Dd11e49D2D",
-    "0x697a70779C1A03Ba2BD28b7627a902BFf831b616",
-    "0x69cDA960E3b20DFD480866fFfd377Ebe40bd0A46",
-    "0x84064c058F2EFea4AB648bB6Bd7e40f83fFDe39a",
-    "0xAF8fD83cFCbe963211FAaf1847F0F217F80B4719",
-    "0xCEb7Da150d16aCE58F090754feF2775C23C8b631",
-  ],
-  opmainnet: [
-    "0x1C9406ee95B7af55F005996947b19F91B6D55b15",
-    "0x37ac9731B0B02df54975cd0c7240e0977a051721",
-    "0x66d5AE25731Ce99D46770745385e662C8e0B4025",
-    "0x6b846E3418455804C1920fA4CC7a31A51C659A2D",
-    "0x9EfdCfC2373f81D3DF24647B1c46e15268884c46",
-  ],
-  unichainmainnet: [
-    "0x0170398083eb0D0387709523baFCA6426146C218",
-    "0x67716D6Bf76170Af816F5735e14c4d44D0B05eD2",
-    "0x68e2A6F7257FAc2F5a557b9E83E1fE6D5B408CE5",
-    "0xB953f92B9f759d97d2F2Dec10A8A3cf75fcE3A95",
-    "0xDa7Ce7Ba016d266645712e2e4Ebc6cC75eA8E4CD",
-    "0xbEC19Bef402C697a7be315d3e59E5F65b89Fa1BB",
-    "0xc219BC179C7cDb37eACB03f993f9fDc2495e3374",
-  ],
-  opbnbmainnet: [
-    "0x13B492B8A03d072Bab5C54AC91Dba5b830a50917",
-    "0x509e81eF638D489936FA85BC58F52Df01190d26C",
-    "0x53d11cB8A0e5320Cd7229C3acc80d1A0707F2672",
-    "0xED827b80Bd838192EA95002C01B5c6dA8354219a",
-    "0xb7a01Ba126830692238521a1aA7E7A7509410b8e",
-  ],
-};
-
-// The single wildcard grant: syncCash() on address(0) → NormalTimelock.
-export const syncCashGrants = (chain: RemoteChain): Permission[] => {
-  const normalTimelock = (NETWORK_ADDRESSES[chain] as Record<string, string>).NORMAL_TIMELOCK;
-  return [{ contractAddress: ZERO, functionSig: SYNC_CASH_SIG, account: normalTimelock }];
-};
-
-// The per-market revokes: syncCash() → NormalTimelock, one per market.
-export const syncCashRevokes = (chain: RemoteChain): Permission[] => {
-  const normalTimelock = (NETWORK_ADDRESSES[chain] as Record<string, string>).NORMAL_TIMELOCK;
-  return SYNC_CASH_MARKETS[chain].map(market => ({
-    contractAddress: market,
-    functionSig: SYNC_CASH_SIG,
-    account: normalTimelock,
-  }));
-};
